refactor(gui): extract Info-tab behavior into InfoPageMixin
Move the Info-tab orchestration off WinpodxWindow into a private mixin
so main_window.py can focus on overall window orchestration:

- InfoPageMixin (src/winpodx/gui/_main_window_info.py)
  Card scaffolding (_info_card), health rendering (_render_health_card,
  _HEALTH_BADGE_COLORS), row-replace helper (_set_info_card_rows),
  gather_info worker orchestration (_refresh_info, _on_info_done,
  _apply_info_snapshot), and auto-refresh timer control
  (_start_info_auto_refresh, _stop_info_auto_refresh).

WinpodxWindow now reads
`class WinpodxWindow(InfoPageMixin, LogsMixin, MaintenanceMixin, QMainWindow)`.

No behavior change; the mixin owns the _info_busy / _info_thread /
_info_worker / _info_auto_timer state lazily. Tests: 1096 passed.

main_window.py: 2255 -> 2016 lines (-239).

Follow-up to the Logs / Maintenance slice. Subsequent slices will
pull out the _build_* UI builders (the biggest remaining chunk),
plus app CRUD and the pod-status timer.
